vga controller impl. done

+++ backend/src/main/java/org/sq5nry/plaszczka/backend/api/vga/IfAmp.java
@@ -9,42 +9,39 @@ public interface IfAmp {
      * Usually a relatively fast decay speed is used. With the speed specified by Vsph,
      * the Vslow control voltage is decayed until it equals Vfast again and the AGC re-enters HANG state.
      *
-     * @param speed 1..255
      *  Reference Voltage: 	5.0V
      *  Minimum Value: 	1 	    19.6mV, 6.06dB/s, 21.6s/131dB
      *  Maximum Value: 	255 	5.0V,   1546dB/s, 84.8ms/131dB
      *  Step Size: 	            19.6mV 	6.06dB/s
      *  Recommended Default Value: 	32 	194dB/s, 0.675s/131dB
      */
-    void setDecaySpeedInDecayStateForHangMode(int speed) throws Exception;
+    void setDecaySpeedInDecayStateForHangMode(float speed) throws Exception;
 
     /**
      * Vspa sets the decay speed for the AGC in Attack/Decay-Mode.
      * Usually a relatively slow decay speed is used for quiet operation of the AGC. With the speed specified by Vspa,
      * the Vslow control voltage is decayed until it equals Vfast again.
      *
-     * @param speed 1..255
      *  Reference Voltage: 	5.0V
      *  Minimum Value: 	    1 	    19.6mV, 6.06dB/s, 21.6s/131dB
      *  Maximum Value: 	    255 	5.0V, 1546dB/s, 84.8ms/131dB
      *  Step Size: 	                19.6mV 	6.06dB/s
      *  Recommended Default Value: 	8 	48.5dB/s, 2.7s/131dB
      */
-    void setDecaySpeedForAttackDecayMode(int speed) throws Exception;
+    void setDecaySpeedForAttackDecayMode(float speed) throws Exception;
 
     /**
      * Vleak sets the decay speed during the HANG state in Hang-Mode. The speed can be 0 but usually a little
      * amount of leaking is configured to have the AGC follow slow QSB during HANG state in Hang-Mode.
      * A voltage divider with buffer is used to provide a more fine grained and smaller output value.
      *
-     * @param speed 0..255
      *  Reference Voltage: 	5.0V
      *  Minimum Value: 	    0 	    0V, 0dB/s
      *  Maximum Value: 	    255 	0.455V, 140.7dB/s
      *  Step Size: 	                1.78mV 	0.551dB/s
      *  Recommended Default Value: 	4 	7.14mV, 2.21dB/s
      */
-    void setDecaySpeedInHangStateForHangMode(int speed) throws Exception;
+    void setDecaySpeedInHangStateForHangMode(float speed) throws Exception;
 
     /**
      * Vfloor sets the compensation for the noise floor of the diode detector. Vfloor is subtracted from the output
@@ -63,21 +60,20 @@ public interface IfAmp {
      *  Step Size: 	                2.99mV 	0.1dB
      *  Recommended Default Value: 	151 	0.452V, 15.1dB
      */
-    void setNoiseFloorCompensation(int val) throws Exception;
+    void setNoiseFloorCompensation(float val) throws Exception;
 
     /**
      * Vath sets the threshold value used for automatically choosing the AGC strategy.
      * Signals with Vslow below Vath will use Attack/Decay-Mode. When Vslow is above Vath, Hang-Mode will be used.
      * Given the default value (112) input signals below S9 will use Attack/Decay-Mode and above S9 Hang-Mode.
      *
-     * @param val 1..255
      *  Reference Voltage: 	3.894V
      *  Minimum Value: 	    0 	    0V, -131dBm
      *  Maximum Value: 	    255 	3.894V, 0dBm
      *  Step Size: 	                15.3mV 	0.515dB
      *  Recommended Default Value: 	112 	1.71V, -73.3dBm, S9
      */
-    void setStrategyThreshold(int val) throws Exception;
+    void setStrategyThreshold(float val) throws Exception;
 
     /**
      * Vhth represents the hang threshold value used to set the direction of the hang integrator.
@@ -85,14 +81,13 @@ public interface IfAmp {
      * In other words: Vhth defines the window in dB that the Hang AGC allows Vfast to decrease during HANG state
      * to maintain its preference for HANG state.
      *
-     * @param val 1..255
      *  Reference Voltage: 	3.894V
      *  Minimum Value: 	    1 	    15.3mV, 0.515dB
      *  Maximum Value: 	    255 	3.894V, 131dB
      *  Step Size: 	                15.3mV 	0.515dB
      *  Recommended Default Value: 	20 	0.305V, 10.3dB
      */
-    void setHangThreshold(int val) throws Exception;
+    void setHangThreshold(float val) throws Exception;
 
     /**
      * Vloop has a dual function:
@@ -111,7 +106,7 @@ public interface IfAmp {
      *  Step Size: 	                15.3mV 	0.515dB
      *  Recommended Default Value: 	12 	0.184V, 6.2dB
      */
-    void setVLoop(int val) throws Exception;
+    void setVLoop(float val) throws Exception;
 
     /**
      * Vgain is used to set the maximum allowed gain of the signal path. This control is usually called IF-gain
@@ -124,7 +119,7 @@ public interface IfAmp {
      *  Step Size: 	                15.3mV 	0.515dB
      *  Recommended Default Value: 	0 	0V, 109.6dB gain
      */
-    void setMaximumGain(int gain) throws Exception;
+    void setMaximumGain(float gain) throws Exception;
 
     /**
      * Vspd sets the maximum hang time of the AGC in Hang-Mode.
@@ -139,7 +134,7 @@ public interface IfAmp {
      *  Step Size: 	                15.3mV
      *  Recommended Default Value: 	15 	0V, 2.0s
      */
-    void setMaximumHangTimeInHangMode(int val) throws Exception;
+    void setMaximumHangTimeInHangMode(float val) throws Exception;
 
     /**
      * Attack sets the attack time of the AGC in both Attack/Decay-Mode and Hang-Mode.
@@ -151,7 +146,7 @@ public interface IfAmp {
      *  Step Size: 	            392Ω 	1.3ms
      *  Recommended Default Value: 	10 	0V, 85ms
      */
-    void setAttackTime(int val) throws Exception;
+    void setAttackTime(float val) throws Exception;
 
     /**
      * Vhot is the "hang on transmit" bit. If set, the hang integrator is forced deep in to HANG state quickly.
